Add files via upload
Phishing awareness training is a program designed to educate individuals and organizations about the risks of phishing attacks. This training teaches people how to recognize phishing attempts, avoid falling prey to them, and act quickly in the event of an attack. By increasing awareness of potential phishing scams, individuals are better equipped to protect themselves and their organizations from the damaging consequences of cybercrime.